REFACTOR: userInput 모듈 분리
